work around TIS Interpreter bug

+++ test.c
@@ -5,10 +5,10 @@
 
 int main(void) {
   for (uint8_t n=1; n < 20; n++)
-    printf("%ju\n", (uintmax_t) divide(10000000000000000000U,n));
+    printf("%llu\n", (unsigned long long) divide(10000000000000000000U,n));
 
   for (uint8_t n=1; n < 20; n++)
-    printf("%ju\n", (uintmax_t)divide(10000000000000000000U-1,n));
+    printf("%llu\n", (unsigned long long)divide(10000000000000000000U-1,n));
 
   uint64_t d[] = {
     18446744073709551610u,
@@ -32,10 +32,10 @@ int main(void) {
     10000000000000000000u,
   };
 
-  printf("\n%ju\n", (uintmax_t)d[0]);
+  printf("\n%llu\n", (unsigned long long)d[0]);
   
   for (uint8_t n=1; n < sizeof d / sizeof d[0] + 1; n++) {
     uint64_t i = d[n-1];
-    printf("%-20ju %-20ju\n", (uintmax_t)divide(i,n), (uintmax_t)divide(i-1,n));
+    printf("%-20llu %-20llu\n", (unsigned long long)divide(i,n), (unsigned long long)divide(i-1,n));
   }
 }
